chore/chatbot-test 오류 수정 중

+++ src/main/java/com/meomulm/chat/model/dto/ChatRequest.java
@@ -13,5 +13,5 @@ public class ChatRequest {
     // 사용자가 보낸 메세지
     private String message;
     // 기존 대화방 ID
-    private Long conversationId;
+    private int conversationId;
 }
